tests: disable some unstable tests on arm macos
