+++ clients/cobol-lsp-vscode-extension/src/test/suite/snippet.spec.test.ts
@@ -18,40 +18,35 @@ import * as vscode from "vscode";
 
 suite("Test snippets with different dialects settings", () => {
   suiteSetup(async function () {
-    this.timeout(30000);
     helper.updateConfig("idms.json");
-    this.timeout(30000);
     await helper.activate();
   });
-  let editor: vscode.TextEditor;
-  test("Autocompletion with IDMS dialect", async () => {
-    await helper.showDocument("SNIPPET.cbl");
-    editor = helper.get_editor("SNIPPET.cbl");
-    await helper.sleep(5000);
-    await helper.insertString(editor, new vscode.Position(1, 0), "   COPY");
-    await vscode.commands.executeCommand(
-      "editor.action.triggerSuggest",
-      editor.document.uri,
-    );
-    await helper.sleep(1000);
-    await helper.executeCommandMultipleTimes("selectNextSuggestion", 1);
-    await helper.sleep(1000);
-    await vscode.commands.executeCommand("acceptSelectedSuggestion");
-    await helper.sleep(1000);
-    const text = editor.document.getText();
-    const acceptedLine = text.split("\n")[4];
-    assert.ok(text, "   COPY IDMS idms-entity.");
-    await helper.sleep(1000);
-    helper.closeActiveEditor();
-  })
-    .timeout(20000)
-    .slow(4000);
+
+  test
+    .only("Autocompletion with IDMS dialect", async () => {
+      await helper.showDocument("SNIPPET.cbl");
+      const editor = helper.get_editor("SNIPPET.cbl");
+      await helper.waitFor(() => editor.document.languageId === "cobol");
+      await helper.insertString(editor, new vscode.Position(1, 0), "   COPY");
+      await vscode.commands.executeCommand(
+        "editor.action.triggerSuggest",
+        editor.document.uri,
+      );
+      await helper.executeCommandMultipleTimes("selectNextSuggestion", 1);
+      await vscode.commands.executeCommand("acceptSelectedSuggestion");
+      const text = editor.document.getText();
+      assert.ok(text, "   COPY IDMS idms-entity.");
+      console.log("we are here!");
+      await helper.sleep(1000);
+      await helper.closeActiveEditor();
+    })
+    .timeout(helper.TEST_TIMEOUT);
 
   test("Autocompletion basic dialect", async () => {
     await helper.showDocument("SNIPPET.cbl");
-    editor = helper.get_editor("SNIPPET.cbl");
+    const editor = helper.get_editor("SNIPPET.cbl");
     helper.updateConfig("basic.json");
-    await helper.sleep(5000);
+    await helper.waitFor(() => editor.document.languageId === "cobol");
     await helper.insertString(editor, new vscode.Position(2, 0), "   COPY");
     await vscode.commands.executeCommand(
       "editor.action.triggerSuggest",
@@ -59,22 +54,17 @@ suite("Test snippets with different dialects settings", () => {
     );
     await helper.sleep(1000);
     await helper.executeCommandMultipleTimes("selectNextSuggestion", 1);
-    await helper.sleep(1000);
     await vscode.commands.executeCommand("acceptSelectedSuggestion");
-    await helper.sleep(1000);
     const text = editor.document.getText();
-    const acceptedLine = text.split("\n")[4];
     assert.strictEqual(text, "   COPY");
-    helper.closeActiveEditor();
-  })
-    .timeout(20000)
-    .slow(4000);
+    await helper.closeActiveEditor();
+  }).timeout(helper.TEST_TIMEOUT);
 
   test("TC152058 Autocompletion basic dialect", async () => {
     await helper.showDocument("USER1.cbl");
-    editor = helper.get_editor("USER1.cbl");
+    const editor = helper.get_editor("USER1.cbl");
     helper.updateConfig("basic.json");
-    await helper.sleep(5000);
+    await helper.waitFor(() => editor.document.languageId === "cobol");
     await helper.insertString(
       editor,
       new vscode.Position(40, 0),
@@ -90,16 +80,17 @@ suite("Test snippets with different dialects settings", () => {
     await editor.edit((edit) => edit.replace(editor.selection, "1"));
     await helper.sleep(1000);
     await vscode.commands.executeCommand("jumpToNextSnippetPlaceholder");
-    await editor.edit((edit) => edit.replace(editor.selection, "str"));
     await helper.sleep(1000);
+    await editor.edit((edit) => edit.replace(editor.selection, "str"));
+    await helper.waitFor(() =>
+      editor.document.getText().includes("ADD 1 TO str"),
+    );
     const text = editor.document.getText();
     const acceptedLine = text.split("\n")[40];
     assert.ok(
       acceptedLine.includes("ADD 1 TO str"),
       "Checks auto complete functionality, also with navigation by snippets",
     );
-    helper.closeActiveEditor();
-  })
-    .timeout(20000)
-    .slow(4000);
+    await helper.closeActiveEditor();
+  }).timeout(helper.TEST_TIMEOUT);
 });
