+++ tests/test_files/project/settings/idms.json
@@ -1,4 +1,7 @@
 {
+    "git.enabled": false,
+    "git.path": null,
+    "git.autofetch": false,
     "cobol-lsp.logging.level.root": "DEBUG",
     "cobol-lsp.cpy-manager.profiles": "",
     "cobol-lsp.cpy-manager.paths-local": [
@@ -8,5 +11,5 @@
       ],
     "cobol-lsp.cpy-manager.paths-dsn": [],
     "editor.snippetSuggestions": "top",
-     "cobol-lsp.dialects": ["IDMS"]
+    "cobol-lsp.dialects": ["IDMS"]
 }
